backend/requirements.txt
asgiref==3.2.10
Django==3.1
djangorestframework==3.11.1
django-cors-headers==3.4.0
drf-yasg==1.20.0
ptvsd==4.3.2
requests==2.25.1
python-dotenv==0.19.0
channels==3.0.0
channels_redis==3.2.0
arrow==1.1.1
django-picklefield==3.0.1
blessed==1.18.1
django-q==1.3.9